@@ -9,16 +9,16 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.1",
     "@types/node": "^16.11.26",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "~17.0.39",
+    "@types/react-dom": "~17.0.11",
     "@types/react-redux": "^7.1.22",
-    "formik": "^2.2.9",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-scripts": "5.0.0",
-    "typescript": "^4.5.5",
-    "web-vitals": "^2.1.4"
+    "formik": "~2.2.9",
+    "react": "~18.0.0",
+    "react-dom": "~18.0.0",
+    "react-redux": "~7.2.8",
+    "react-scripts": "~5.0.0",
+    "typescript": "~4.5.5",
+    "web-vitals": "~2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -47,6 +47,6 @@
   "devDependencies": {
     "autoprefixer": "^10.4.2",
     "postcss": "^8.4.7",
-    "tailwindcss": "^3.0.23"
+    "tailwindcss": "~3.0.23"
   }
 }
