@@ -1,14 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import {store} from './store'
 import {Provider} from 'react-redux'
 
-ReactDOM.render(
+const rootElement = document.getElementById('root')
+if (!rootElement) throw new Error('Failed to find the root element')
+const root = ReactDOM.createRoot(rootElement)
+root.render(
     <Provider store={store}>
         <App/>
-    </Provider>,
-    document.getElementById('root')
-);
+    </Provider>
+)
 
